\name{insertBlankLineUnder}
\title{Insert Blank Line Under Current Line}
\usage{
insertBlankLineUnder()
}
\description{
This command will insert a blank line under the current cursor position and move the cursor to the beginning of the line.  This is intended to be used from a hotkey via an addin.
}
\examples{
insertBlankLineUnder()
}
